move header_iter into the test module
